Rename testing project
