+++ src/test-libwfp/stdafx.cpp
