+++ src/test-libwfp/test-libwfp.vcxproj
@@ -183,7 +183,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="tests.cpp" />
+    <ClCompile Include="test-libwfp.cpp" />
     <ClCompile Include="wfp\ipaddress.cpp" />
     <ClCompile Include="wfp\layerconditions.cpp" />
   </ItemGroup>
